refactor: rename components dir to uis
