+++ .storybook/index.ts
@@ -1,3 +1,3 @@
-import '../stories/components';
+import '../stories/uis';
 import '../stories/modals';
 import '../stories/packages';
