+++ docs/.storybook/webpack.config.js
@@ -13,7 +13,7 @@ const ttfLoaderConfig = {
   ],
   include: [
     resolve(__dirname, '../', "node_modules/react-native-vector-icons"),
-    resolve(__dirname, '../', "node_modules/dooboo-ui/components/Icon"),
+    resolve(__dirname, '../', "node_modules/dooboo-ui/ui/Icon"),
   ],
 };
