+++ docs/src/overview/readme.stories.mdx
@@ -103,7 +103,7 @@ module.exports = {
     ios: {},
     android: {},
   },
-+  assets: ['dooboo-ui/components/Icon/doobooui.ttf'],
++  assets: ['dooboo-ui/uis/Icon/doobooui.ttf'],
 };
 ```
 
@@ -119,7 +119,7 @@ module.exports = {
 import {useFonts} from 'expo-font';
 
 const [fontsLoaded] = useFonts({
-  IcoMoon: require('dooboo-ui/components/Icon/doobooui.ttf'),
+  IcoMoon: require('dooboo-ui/uis/Icon/doobooui.ttf'),
 });
 ```
