+++ main/index.ts
@@ -1,19 +1,19 @@
-export * from './components/Typography';
-export * from './components/RadioGroup';
+export * from './uis/Typography';
+export * from './uis/RadioGroup';
 export * from './modals/Snackbar';
-export * from './components/Icon';
-export * from './components/Hr';
-export * from './components/Button';
-export * from './components/ButtonGroup';
-export * from './components/IconButton';
-export * from './components/Checkbox';
-export * from './components/EditText';
-export * from './components/NetworkImage';
-export * from './components/ProgressBar';
-export * from './components/ProgressCircle';
-export * from './components/LoadingIndicator';
-export * from './components/Accordion';
-export * from './components/StatusbarBrightness';
-export * from './components/SwitchToggle';
-export * from './components/FAB';
+export * from './uis/Icon';
+export * from './uis/Hr';
+export * from './uis/Button';
+export * from './uis/ButtonGroup';
+export * from './uis/IconButton';
+export * from './uis/Checkbox';
+export * from './uis/EditText';
+export * from './uis/NetworkImage';
+export * from './uis/ProgressBar';
+export * from './uis/ProgressCircle';
+export * from './uis/LoadingIndicator';
+export * from './uis/Accordion';
+export * from './uis/StatusbarBrightness';
+export * from './uis/SwitchToggle';
+export * from './uis/FAB';
 export * from './providers';
