+++ main/uis/LoadingIndicator/__tests__/LoadingIndicator.test.tsx
