+++ main/uis/StatusbarBrightness/__tests__/StatusBarBrightness.test.tsx
