+++ package.json
@@ -21,7 +21,7 @@
     "test": "yarn test:pre; jest; yarn test:post;",
     "test:all": "yarn pre && yarn lint && yarn tsc && yarn test",
     "build": "tsc --project tsconfigBuild.json && yarn cp:icon",
-    "cp:icon": "mkdir -p `dirname ./lib/components/Icon/selection.json` && cp ./main/components/Icon/selection.json ./lib/components/Icon/selection.json && cp ./main/components/Icon/doobooui.ttf ./lib/components/Icon/doobooui.ttf",
+    "cp:icon": "mkdir -p `dirname ./lib/uis/Icon/selection.json` && cp ./main/uis/Icon/selection.json ./lib/uis/Icon/selection.json && cp ./main/uis/Icon/doobooui.ttf ./lib/uis/Icon/doobooui.ttf",
     "postbuild": "cp ./main/react-native.config.cjs ./lib/react-native.config.cjs && rm -rf ./lib/tsconfig* && cp README.md ./lib/README.md && cp -R main/__assets__ lib",
     "build:storybook": "expo export:web && sh web-build-postinstall.sh",
     "deploy:storybook": "gh-pages -d web-build",
